src/DeepLens.FeatureExtractionService/requirements.txt
# Feature Extraction Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML/AI Dependencies
onnxruntime==1.20.1
pillow==10.1.0
numpy==1.26.2

# Logging and Monitoring
python-json-logger==2.0.7
